requirements_analyzer.txt
# Phoenix LLM Analyzer Requirements
# Core dependencies
arize-phoenix==11.16.1
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Export functionality
openpyxl>=3.1.0

# Optional: For enhanced data processing
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
